Move unit tests to own files
Having the unit tests related to a request in its own file is a good
thing, i think. Having all of the tests in one big file was quickly
going to be an issue.

Also creates a ModelMocks-file that creates mocked instances of our
models. Some unit tests are dependant on having valid model instances.
Since there are only two now they are in the same file. Might move to
separate files in the future if the files becomes to big.