Move credentials to Credentials file in unit tests
The purpose of this is to make it harder for me to accidentally commit
my real credentials to this repo. Unit tests are now always run against
the prod api. If you want to test it, you have to add the missing
Credentials.swift file and add the required functions.